+++ src/Pages/User.jsx
@@ -1,11 +1,13 @@
-import { Box, Button, Checkbox, Heading, Input, InputGroup, InputLeftElement, Table, TableContainer, Tbody, Th, Thead, Tr } from '@chakra-ui/react';
+import { Box, Button, Heading, Input, InputGroup, InputLeftElement, Table, TableContainer, Tbody, Th, Thead, Tr } from '@chakra-ui/react';
 import axios from 'axios';
 import React from 'react';
 import { MdOutlineAdd, MdSearch } from "react-icons/md"
+import { useNavigate } from 'react-router-dom';
 import TableUser from '../Components/TableUser';
 import { API_URL } from '../helper';
 
 function NewUser() {
+    const navigate = useNavigate()
 
     const [userList, setUserList] = React.useState([]);
 
@@ -28,12 +30,10 @@ function NewUser() {
     const printUser = () => {
         console.log(`ini dari userlist`, userList);
         let newArr = userList.map((val, idx) => {
-            return <TableUser name={val.name} role={val.role.role} phone={val.phone} email={val.email} />
+            return <TableUser name={val.name} role={val.role.role} phone={val.phone} email={val.email} uuid={val.uu_id} getAllUser={getAllUser} />
         })
         return newArr
     }
-
-
 
     return (
         <Box m={"40px"}>
@@ -41,9 +41,10 @@ function NewUser() {
                 User List
             </Heading>
 
-            <Button mt={"28px"} _hover={"none"} bgColor={"#537FE7"} style={{ color: "#E9F8F9" }} leftIcon={<MdOutlineAdd />}> Add New User</Button>
+            <Button onClick={() => navigate("/addnewuser")} mt={"28px"} _hover={"none"} bgColor={"#537FE7"} style={{ color: "#E9F8F9" }} leftIcon={<MdOutlineAdd />}>
+                Add New User</Button>
 
-            <InputGroup mt={"28px"} w="96">
+            <InputGroup mt={"28px"} w={{ sm: "40", md: "96", lg: "96" }}>
                 <InputLeftElement pointerEvents="none" children={<MdSearch size={"22"} color='gray.800' />} />
                 <Input type="text" placeholder="Search List" bg="white" color="gray.800" />
             </InputGroup>
@@ -52,7 +53,6 @@ function NewUser() {
                 <Table >
                     <Thead>
                         <Tr>
-                            <Th children={<Checkbox />}></Th>
                             <Th>Name</Th>
                             <Th>Role</Th>
                             <Th>Phone</Th>
@@ -65,12 +65,7 @@ function NewUser() {
 
                 </Table>
             </TableContainer>
-
-
-
         </Box>
-
-
     );
 }
